requirements.txt
Flask==2.1.2                  # The main web framework
Flask-JWT-Extended==4.4.4     # For JWT-based authorization
Flask-Cors==3.0.10            # To handle CORS (Cross-Origin Resource Sharing) if needed
pandas==1.4.2                 # For easier CSV handling and calculations
pytest==7.1.2                 # For writing unit tests
pytest-flask==1.2.0           # For testing Flask applications with pytest
SQLAlchemy==1.4.36            # Provides an easy way to interact with the SQLite database using Python objects instead of raw SQL
flask_sqlalchemy==2.5.1       # An extension for Flask that adds support for SQLAlchemy, simplifying database interactions within Flask applications
Werkzeug==2.1.2               # A comprehensive WSGI web application library